Add a cast to i64 for an integer in an sqlite3OsWrite() offset calculation.  The cast is not strictly necessary, but it helps human readers see that the code is correct.
FossilOrigin-Name: 7564ff1ba2c2fba89106d1aa06cc5379e752f119f22370f2f155f24cc698dec6

+++ src/pager.c
@@ -6593,7 +6593,7 @@ int sqlite3PagerCommitPhaseOne(
             int szPage = (int)pPager->pageSize;
             memset(pTmp, 0, szPage);
             rc = sqlite3OsWrite(pPager->fd, pTmp, szPage,
-                      (pPager->dbSize*pPager->pageSize)-szPage);
+                      ((i64)pPager->dbSize*pPager->pageSize)-szPage);
           }
           if( rc==SQLITE_OK ){
             rc = sqlite3OsFileControl(fd, SQLITE_FCNTL_COMMIT_ATOMIC_WRITE, 0);
